refactor(middleware-files): rename middleware files
