+++ server/app/middleware/document.js
